WorkBook: Add a datastructure Detach. This allows to detach a sheet from a workbook while retaining its index in the workbook. This helps when modifying Sheets as this often means modifying the WorkBook itself too. And this immediately crashes with 2 mutable borrows.
The detached value comes in the form of another smart pointer, but one that
completely forgets where it came from. The only thing it remembers is the original
index (key) of whatever datastructure it was stored in. The datastructure retains
an empty slot where the value was, and has to cope with this fact. Which means
it panics when a detached value should be accessed.

Reattaching the value works quite simple with this, as the correct index (key) is
already known.

Sheet:
The colwidth is now stored with the ColumnHeader, so its much easier to set it.

The col_width is copied from the style into the column header and vice versa when writing.
**This is a breaking change as write_ods now needs a &mut WorkBook.***
But this is not the only such thing that I plan to do, so be it.

Length gets a new entry Length::Default. The intention is "use whatever", or "use optimal". Tried a second Length enum, but that was not very nice. Plays nice with
style:use-optimal-column-width, maybe that's a hint that this is ok.

ColStyle uses the new Length::Default.