refactor: rebuild the clone logic.
rebuild cell-cloning in a single postprocessing step.

feat: split RowHeader.repeat in two fields repeat and span.

Repeat is used to read the repeat-value and moved to span after cloning the cells. This way cloning the row-headers can be avoided. after this point span indicates the row-range where the header-values are valid. access-methods use this to find the correct row-header. uses copy-on-write whenever a value is set.

the now free repeat value can be used for application code to generate whatever row-duplication it wants.